Move cap URL to a source plugin
